+++ mbed-hal/spi_api.h
@@ -32,14 +32,19 @@
 #define SPI_FILL_WORD         (0xFFFF)
 
 #if DEVICE_SPI_ASYNCH
+/** Asynch spi hal structure
+ */
 typedef struct {
-    struct spi_s spi;
-    struct buffer_s tx_buff;
-    struct buffer_s rx_buff;
+    struct spi_s spi;        /**< Target specific spi structure */
+    struct buffer_s tx_buff; /**< Tx buffer */
+    struct buffer_s rx_buff; /**< Rx buffer */
 } spi_t;
 
 #else
+/** Non-asynch spi hal structure
+ */
 typedef struct spi_s spi_t;
+
 #endif
 
 #ifdef __cplusplus
@@ -154,7 +159,7 @@ uint8_t spi_get_module(spi_t *obj);
 
 /** Begin the SPI transfer. Buffer pointers and lengths are specified in tx_buff and rx_buff
  *
- * @param[in] obj     The SPI object which holds the transfer information
+ * @param[in] obj       The SPI object which holds the transfer information
  * @param[in] tx        The buffer to send
  * @param[in] tx_length The number of words to transmit
  * @param[in] rx        The buffer to receive
@@ -164,7 +169,7 @@ uint8_t spi_get_module(spi_t *obj);
  * @param[in] handler   SPI interrupt handler
  * @param[in] hint      A suggestion for how to use DMA with this transfer
  */
-void spi_master_transfer(spi_t *obj, void *tx, uint32_t tx_length, void *rx, uint32_t rx_length, uint8_t bit_width, uint32_t handler, uint32_t event, DMAUsage hint);
+void spi_master_transfer(spi_t *obj, void *tx, size_t tx_length, void *rx, size_t rx_length, uint8_t bit_width, uint32_t handler, uint32_t event, DMAUsage hint);
 
 /** The asynchronous IRQ handler
  *
